+++ CHANGES.rst
@@ -10,6 +10,8 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 5.11.2 (unreleased)
 -------------------
 
+- Fix error messages from spam/pen test requests.
+
 - Fix a ``ResourceWarning`` emitted when uploading large files.
   (`#1242 <https://github.com/zopefoundation/Zope/issues/1242>`_)
 
@@ -23,6 +25,7 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
   (`#1239 <https://github.com/zopefoundation/Zope/issues/1239>`_).
 
 - Update to ``zope.interface = 7.2``.
+
 
 5.11.1 (2024-11-03)
 -------------------